requirements-common.txt
# ── Core Framework ───────────────────────────────────────
streamlit>=1.32.0

# ── LLM APIs ────────────────────────────────────────────
# Gemini API is accessed via the stdlib REST helper in shared/llm.py

# ── Data Processing ──────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ── Visualization ────────────────────────────────────────
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── NLP ─────────────────────────────────────────────────
spacy>=3.7.0
vaderSentiment>=3.3.2
scikit-learn>=1.4.0

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0

# ── Dev / Testing ────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
